% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufn_rq.R
\docType{methods}
\name{ApplyTo,rq_u_fn_w,data.frame-method}
\alias{ApplyTo,rq_u_fn_w,data.frame-method}
\alias{ApplyTo,rq_u_fn_w,UnaryFnList-method}
\alias{ApplyTo,rq_u_fn_w,UnaryFn-method}
\title{Apply a single argument function to its argument.}
\usage{
\S4method{ApplyTo}{rq_u_fn_w,data.frame}(f, x, env = parent.frame())

\S4method{ApplyTo}{rq_u_fn_w,UnaryFnList}(f, x, env = parent.frame())

\S4method{ApplyTo}{rq_u_fn_w,UnaryFn}(f, x, env = parent.frame())
}
\arguments{
\item{f}{object of S4 class derived from UnaryFn.}

\item{x}{argument.}

\item{env}{environment to work in.}
}
\value{
f(x) if x is not a UnaryFn else f composed with x.
}
\description{
Apply a single argument function to its argument.
}
