% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_drop_columns.R
\name{ex_data_table.relop_drop_columns}
\alias{ex_data_table.relop_drop_columns}
\title{Implement drop columns.}
\usage{
\method{ex_data_table}{relop_drop_columns}(
  optree,
  ...,
  tables = list(),
  source_limit = NULL,
  source_usage = NULL,
  env = parent.frame()
)
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_limit}{if not null limit all table sources to no more than this many rows (used for debugging).}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{env}{environment to work in.}
}
\description{
\code{data.table} based implementation.
}
\examples{

dL <- data.frame(x = 1, y = 2, z = 3)
rquery_pipeline <- local_td(dL) \%.>\%
  drop_columns(., "y")
ex_data_table(rquery_pipeline)

}
