% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_relop_list.R
\name{ex_data_table.relop_list}
\alias{ex_data_table.relop_list}
\title{ex_data_table for relop_list.}
\usage{
\method{ex_data_table}{relop_list}(optree, ..., tables = list(),
  source_usage = NULL, source_limit = NULL, env = parent.frame())
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{source_limit}{if not null limit all table sources to no more than this many rows (used for debugging).}

\item{env}{environment to work in.}
}
\description{
Execute storing intermeidate tables in tables variable.
}
