% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_non_sql.R
\name{ex_data_table_funciton_node}
\alias{ex_data_table_funciton_node}
\title{Helper to build data.table capable non-sql nodes.}
\usage{
ex_data_table_funciton_node(., f, ..., columns_produced, display_form)
}
\arguments{
\item{.}{or data.frame input.}

\item{f}{function that takes a data.table to a data.frame (or data.table).}

\item{...}{force later arguments to bind by name.}

\item{columns_produced}{character columns produces by f.}

\item{display_form}{display form for node.}
}
\value{
relop non-sql node implementation.
}
\description{
Helper to build data.table capable non-sql nodes.
}
\seealso{
\code{\link[rqdatatable]{ex_data_table.relop_non_sql}}
}
