% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery_fns.R
\name{ex_data_table_step.default}
\alias{ex_data_table_step.default}
\title{default non-impementation.}
\usage{
\method{ex_data_table_step}{default}(
  optree,
  ...,
  tables = list(),
  source_usage = NULL,
  source_limit = NULL,
  env = parent.frame()
)
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{source_limit}{if not null limit all table sources to no more than this many rows (used for debugging).}

\item{env}{environment to work in.}
}
\description{
Throw on error if this method is called, signalling that a specific \code{data.table} implemetation is needed for this method.
}
