% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_order_expr.R
\name{ex_data_table_step.relop_order_expr}
\alias{ex_data_table_step.relop_order_expr}
\title{Order rows by expression.}
\usage{
\method{ex_data_table_step}{relop_order_expr}(
  optree,
  ...,
  tables = list(),
  source_usage = NULL,
  source_limit = NULL,
  env = parent.frame()
)
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{source_limit}{if not null limit all table sources to no more than this many rows (used for debugging).}

\item{env}{environment to work in.}
}
\description{
\code{data.table} based implementation.
}
\examples{

dL <- build_frame(
    "x", "y" |
    2L , "b" |
   -4L , "a" |
    3L , "c" )
rquery_pipeline <- local_td(dL) \%.>\%
  order_expr(., abs(x))
dL \%.>\% rquery_pipeline

}
