% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufn_rq.R
\name{format.rq_u_fn_w}
\alias{format.rq_u_fn_w}
\title{format step}
\usage{
\method{format}{rq_u_fn_w}(x, ...)
}
\arguments{
\item{x}{object to format}

\item{...}{additional aguments (not used)}
}
\value{
character
}
\description{
format step
}
