% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_exports.R
\name{rbindlist_data_table}
\alias{rbindlist_data_table}
\title{rbindlist}
\usage{
rbindlist_data_table(l, use.names = TRUE, fill = TRUE, idcol = NULL)
}
\arguments{
\item{l}{list of data.frames to rbind.}

\item{use.names}{passed to data.table}

\item{fill}{passed to data.table}

\item{idcol}{passed to data.table}
}
\value{
data.table
}
\description{
Note: different argument defaults than data.table::rbindlist.
}
\examples{

rbindlist_data_table(list(
  data.frame(x = 1, y = 2),
  data.frame(x = c(2, 3), y = c(NA, 4))))

}
\keyword{internal}
