% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufn_rq.R
\name{rq_fn_wrapper}
\alias{rq_fn_wrapper}
\title{Wrap an rquery pipeline exprssion as a function.}
\usage{
rq_fn_wrapper(rqpipeline)
}
\arguments{
\item{rqpipeline}{rquery pipeline.}
}
\value{
rq_u_fn_w
}
\description{
Wrap an rquery pipeline exprssion as a function.
}
