% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq_partial.R
\name{rq_partial}
\alias{rq_partial}
\title{Wrap a function name and arguments as an rqdatatable stage.}
\usage{
rq_partial(source, fn_name = NULL, ..., fn_package = "base",
  arg_name = "", args = list(), columns_produced,
  env = parent.frame())
}
\arguments{
\item{source}{incoming relop pipeline.}

\item{fn_name}{character, name of function.}

\item{...}{force later arguments to be taken by name.}

\item{fn_package}{character, name of package.}

\item{arg_name}{name for remaining argument.}

\item{args}{list of function argument values}

\item{columns_produced}{columns of this node's result.}

\item{env}{environment to work in.}
}
\value{
wrapped function
}
\description{
Wrap a function name and arguments as an rqdatatable stage.
}
\seealso{
\code{\link{applyto}}
}
