% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq_partial.R
\name{rq_partialf}
\alias{rq_partialf}
\title{Wrap a function name and arguments as an rqdatatable stage.}
\usage{
rq_partialf(source, fn = NULL, ..., arg_name = "", args = list(),
  columns_produced = NULL, check_result_details = TRUE,
  env = parent.frame())
}
\arguments{
\item{source}{incoming relop pipeline.}

\item{fn}{a function to wrap.}

\item{...}{force later arguments to be taken by name.}

\item{arg_name}{name for remaining argument.}

\item{args}{list of function argument values}

\item{columns_produced}{columns of this node's result.}

\item{check_result_details}{logical, if TRUE enforce result type and columns.}

\item{env}{environment to work in.}
}
\value{
wrapped function
}
\description{
Wrap a function name and arguments as an rqdatatable stage.
}
\seealso{
\code{\link{applyto}}
}
