% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rqdatatable-package}
\alias{rqdatatable}
\alias{rqdatatable-package}
\title{\code{rqdatatable}: Relational Query Generator for Data Manipulation Implemented by data.table}
\description{
Implements the \code{rquery} piped query algebra using \code{data.table}.  This allows
for a high-speed in memory implementation of Codd-style data manipulation tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WinVector/rqdatatable/}
  \item \url{https://winvector.github.io/rqdatatable/}
  \item Report bugs at \url{https://github.com/WinVector/rqdatatable/issues}
}

}
\author{
\strong{Maintainer}: John Mount \email{jmount@win-vector.com}

Other contributors:
\itemize{
  \item  Win-Vector LLC [copyright holder]
}

}
