% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rqdatatable}
\alias{rqdatatable}
\title{\code{rqdatatable}: Relational Query Generator for Data Manipulation Implemented by data.table}
\description{
Implements the \code{rquery} piped query algebra using \code{data.table}.  This allows
for a high-speed in memory implementation of Codd-style data manipulation tools.
}
