% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_rqdatatable_as_executor}
\alias{set_rqdatatable_as_executor}
\title{Set rqdatatable package as default rquery executor}
\usage{
set_rqdatatable_as_executor()
}
\description{
Sets rqdatatable (and hence data.table) as the default executor for rquery).
}
