% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufn_rq.R
\docType{methods}
\name{show,rq_u_fn_w-method}
\alias{show,rq_u_fn_w-method}
\title{S4 print method}
\usage{
\S4method{show}{rq_u_fn_w}(object)
}
\arguments{
\item{object}{item to print}
}
\description{
S4 print method
}
