% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_natural_join.R
\name{ex_data_table.relop_natural_join}
\alias{ex_data_table.relop_natural_join}
\title{Natural join.}
\usage{
\method{ex_data_table}{relop_natural_join}(optree, ..., tables = list(),
  source_usage = NULL, source_limit = NULL, env = parent.frame())
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{source_limit}{if not null limit all table sources to no more than this many rows (used for debugging).}

\item{env}{environment to work in.}
}
\description{
\code{data.table} based implementation.  Note: does not yet match \code{rquery} common column coalesce semantics,
so all common columns must be join conditions.
}
\examples{

d1 <- build_frame(
    "key", "val", "val1" |
      "a"  , 1  ,  10    |
      "b"  , 2  ,  11    |
      "c"  , 3  ,  12    )
d2 <- build_frame(
    "key", "val", "val2" |
      "a"  , 5  ,  13    |
      "b"  , 6  ,  14    |
      "d"  , 7  ,  15    )
# can't have shared non-key columns in rqdatatable yet
d1$val <- NULL
d2$val <- NULL

# key matching join
optree <- natural_join(local_td(d1), local_td(d2),
                       jointype = "FULL", by = 'key')
ex_data_table(optree)[]

# can't have shared non-key columns in rqdatatable yet
# # full cross-product join
# optree2 <- natural_join(local_td(d1), local_td(d2),
#                         jointype = "FULL", by = NULL)
# ex_data_table(optree2)[]

}
