% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_natural_join.R
\name{ex_data_table.relop_natural_join}
\alias{ex_data_table.relop_natural_join}
\title{Natural join.}
\usage{
\method{ex_data_table}{relop_natural_join}(optree, ..., tables = list(),
  source_usage = NULL, env = parent.frame())
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{env}{environment to work in.}
}
\description{
\code{data.table} based implementation.  Note: does not yet match \code{rquery} common column coalesce semantics,
so all common columns must be join conditions.
}
\examples{

a <- build_frame(
    "x", "y" |
    1L , "a" |
    2L , "b" )
b <- build_frame(
    "x", "z" |
    2L , "x" |
    3L , "y" )
rquery_pipeline <-
  natural_join(local_td(a), local_td(b),
               by = "x",
               jointype = "FULL")
ex_data_table(rquery_pipeline)[]

}
