% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relop_non_sql.R
\name{ex_data_table.relop_non_sql}
\alias{ex_data_table.relop_non_sql}
\title{Direct non-sql (function) node, not implented for \code{data.table} case.}
\usage{
\method{ex_data_table}{relop_non_sql}(optree, ..., tables = list(),
  source_usage = NULL, env = parent.frame())
}
\arguments{
\item{optree}{relop operations tree.}

\item{...}{not used, force later arguments to bind by name.}

\item{tables}{named list map from table names used in nodes to data.tables and data.frames.}

\item{source_usage}{list mapping source table names to vectors of columns used.}

\item{env}{environment to work in.}
}
\description{
Placeholder to through legible "not implemented" message.
}
