% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp_rquery.R, R/rquery_fns.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{assign_slice}
\alias{column_names}
\alias{columns_used}
\alias{commencify}
\alias{complete_design}
\alias{count_null_cols}
\alias{db_td}
\alias{drop_columns}
\alias{execute}
\alias{extend_nse}
\alias{extend_se}
\alias{format_node}
\alias{if_else_block}
\alias{if_else_op}
\alias{local_td}
\alias{map_column_values}
\alias{mark_null_cols}
\alias{mk_td}
\alias{natural_join}
\alias{non_sql_node}
\alias{normalize_cols}
\alias{null_replace}
\alias{op_diagram}
\alias{orderby}
\alias{pick_top_k}
\alias{project_nse}
\alias{project_se}
\alias{rename_columns}
\alias{rquery_apply_to_data_frame}
\alias{select_columns}
\alias{select_rows_nse}
\alias{select_rows_se}
\alias{set_indicator}
\alias{table_source}
\alias{tables_used}
\alias{theta_join_nse}
\alias{theta_join_se}
\alias{unionall}
\alias{mk_tmp_name_source}
\alias{map_to_char}
\alias{\%.>\%}
\alias{\%:=\%}
\alias{let}
\alias{qc}
\alias{qe}
\alias{build_frame}
\alias{draw_frame}
\alias{qchar_frame}
\alias{apply_left}
\alias{apply_right}
\alias{as.data.table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{rquery}{\code{\link[rquery]{assign_slice}}, \code{\link[rquery]{column_names}}, \code{\link[rquery]{columns_used}}, \code{\link[rquery]{commencify}}, \code{\link[rquery]{complete_design}}, \code{\link[rquery]{count_null_cols}}, \code{\link[rquery]{db_td}}, \code{\link[rquery]{drop_columns}}, \code{\link[rquery]{execute}}, \code{\link[rquery]{extend_nse}}, \code{\link[rquery]{extend_se}}, \code{\link[rquery]{format_node}}, \code{\link[rquery]{if_else_block}}, \code{\link[rquery]{if_else_op}}, \code{\link[rquery]{local_td}}, \code{\link[rquery]{map_column_values}}, \code{\link[rquery]{mark_null_cols}}, \code{\link[rquery]{mk_td}}, \code{\link[rquery]{natural_join}}, \code{\link[rquery]{non_sql_node}}, \code{\link[rquery]{normalize_cols}}, \code{\link[rquery]{null_replace}}, \code{\link[rquery]{op_diagram}}, \code{\link[rquery]{orderby}}, \code{\link[rquery]{pick_top_k}}, \code{\link[rquery]{project_nse}}, \code{\link[rquery]{project_se}}, \code{\link[rquery]{rename_columns}}, \code{\link[rquery]{rquery_apply_to_data_frame}}, \code{\link[rquery]{select_columns}}, \code{\link[rquery]{select_rows_nse}}, \code{\link[rquery]{select_rows_se}}, \code{\link[rquery]{set_indicator}}, \code{\link[rquery]{table_source}}, \code{\link[rquery]{tables_used}}, \code{\link[rquery]{theta_join_nse}}, \code{\link[rquery]{theta_join_se}}, \code{\link[rquery]{unionall}}}

  \item{wrapr}{\code{\link[wrapr]{\%:=\%}}, \code{\link[wrapr]{\%.>\%}}, \code{\link[wrapr]{apply_left}}, \code{\link[wrapr]{apply_right}}, \code{\link[wrapr]{build_frame}}, \code{\link[wrapr]{draw_frame}}, \code{\link[wrapr]{let}}, \code{\link[wrapr]{map_to_char}}, \code{\link[wrapr]{mk_tmp_name_source}}, \code{\link[wrapr]{qc}}, \code{\link[wrapr]{qc}}, \code{\link[wrapr]{qchar_frame}}, \code{\link[wrapr]{qe}}}
}}

