% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq_partial.R
\name{rq_partial}
\alias{rq_partial}
\title{Wrap a function name and arguments as an rqdatatable stage.}
\usage{
rq_partial(source, step, ..., columns_produced = NULL,
  check_result_details = TRUE, env = parent.frame())
}
\arguments{
\item{source}{incoming relop pipeline.}

\item{step}{a wrapr::UnaryFn derived S4 instance.}

\item{...}{force later arguments to be taken by name.}

\item{columns_produced}{columns of this node's result.}

\item{check_result_details}{logical, if TRUE enforce result type and columns.}

\item{env}{environment to work in.}
}
\value{
wrapped function
}
\description{
Wrap a function name and arguments as an rqdatatable stage.
}
\seealso{
\code{\link{applyto}}
}
