% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq_ufn.R
\name{rq_ufn}
\alias{rq_ufn}
\title{Wrap a function name and arguments as an rqdatatable stage.}
\usage{
rq_ufn(source, step, ..., columns_produced = NULL,
  check_result_details = TRUE, use_data_table = TRUE, f_db = NULL,
  temp_name_source = wrapr::mk_tmp_name_source(), env = parent.frame())
}
\arguments{
\item{source}{incoming relop pipeline.}

\item{step}{a wrapr::UnaryFn derived S4 instance.}

\item{...}{force later arguments to be taken by name.}

\item{columns_produced}{columns of this node's result.}

\item{check_result_details}{logical, if TRUE enforce result type and columns.}

\item{use_data_table}{logical, if TRUE use data.table code path.}

\item{f_db}{database implementation signature: f_db(db, incoming_table_name, outgoing_table_name, nd) (db being a database handle, can't be a nested rquery pipeline).}

\item{temp_name_source}{a wrapr::mk_tmp_name_source().}

\item{env}{environment to work in.}
}
\value{
wrapped function
}
\description{
Wrap a function name and arguments as an rqdatatable stage.
}
